Update Configuration_adv.h
USE_WATCHDOG should be on by default. Safe operation is quite important. I know the Arduino bootloader fails when the watchdog triggers, but this is safer operation then having a run-away heater. If you manage to trigger the watchdog by accident you will have a bad printing experience anyhow, due to the fact that the watchdog has quite a long timeout (seconds), and thus if the watchdog does not get triggered, the planner is also most likely stopped.

Result of printer almost burning down a house (again) discussion:
https://plus.google.com/+AnthonyMorris/posts/LsuG1cJXo6Y